+++ test/finufft3d_test.cpp
@@ -6,7 +6,7 @@
 #include <iostream>
 #include <iomanip>
 
-// how big a problem to do full direct DFT check...
+// how big a problem to do full direct DFT check in 3D...
 #define BIGPROB 1e8
 
 int main(int argc, char* argv[])
@@ -22,18 +22,18 @@ int main(int argc, char* argv[])
    Barnett 2/2/17
 */
 {
-  BIGINT M = 1e6, N1 = 100, N2 = 200, N3 = 50;  // defaults: M = # srcs, N1,N2,N3 = # modes
+  INT M = 1e6, N1 = 100, N2 = 200, N3 = 50;  // defaults: M = # srcs, N1,N2,N3 = # modes
   double w, tol = 1e-6;       // default
   nufft_opts opts;
   opts.debug = 0;             // 1 to see some timings
   opts.spread_sort = 1;       // default
   int isign = +1;             // choose which exponential sign to test
   if (argc>1) {
-    sscanf(argv[1],"%lf",&w); N1 = (BIGINT)w;
-    sscanf(argv[2],"%lf",&w); N2 = (BIGINT)w;
-    sscanf(argv[3],"%lf",&w); N3 = (BIGINT)w;
+    sscanf(argv[1],"%lf",&w); N1 = (INT)w;
+    sscanf(argv[2],"%lf",&w); N2 = (INT)w;
+    sscanf(argv[3],"%lf",&w); N3 = (INT)w;
   }
-  if (argc>4) { sscanf(argv[4],"%lf",&w); M = (BIGINT)w; }
+  if (argc>4) { sscanf(argv[4],"%lf",&w); M = (INT)w; }
   if (argc>5) {
     sscanf(argv[5],"%lf",&tol);
     if (tol<=0.0) { printf("tol must be positive!\n"); return 1; }
@@ -46,12 +46,12 @@ int main(int argc, char* argv[])
     return 1;
   }
   cout << scientific << setprecision(15);
-  BIGINT N = N1*N2*N3;
+  INT N = N1*N2*N3;
 
   double *x = (double *)malloc(sizeof(double)*M);        // NU pts x coords
   double *y = (double *)malloc(sizeof(double)*M);        // NU pts y coords
   double *z = (double *)malloc(sizeof(double)*M);        // NU pts z coords
-  for (BIGINT j=0; j<M; ++j) {
+  for (INT j=0; j<M; ++j) {
     x[j] = M_PI*randm11();
     y[j] = M_PI*randm11();
     z[j] = M_PI*randm11();
@@ -60,7 +60,7 @@ int main(int argc, char* argv[])
   dcomplex* F = (dcomplex*)malloc(sizeof(dcomplex)*N);   // mode ampls
 
   printf("test 3d type-1:\n"); // -------------- type 1
-  for (BIGINT j=0; j<M; ++j) c[j] = crandm11();
+  for (INT j=0; j<M; ++j) c[j] = crandm11();
   CNTime timer; timer.start();
   int ier = finufft3d1(M,x,y,z,c,isign,tol,N1,N2,N3,F,opts);
   double ti=timer.elapsedsec();
@@ -69,16 +69,16 @@ int main(int argc, char* argv[])
     exit(ier);
   } else
     printf("     %ld NU pts to (%ld,%ld,%ld) modes in %.3g s \t%.3g NU pts/s\n",
-	   M,N1,N2,N3,ti,M/ti);
+	   (INT64)M,(INT64)N1,(INT64)N2,(INT64)N3,ti,M/ti);
 
-  BIGINT nt1 = (BIGINT)(0.37*N1), nt2 = (BIGINT)(0.26*N2), nt3 = (BIGINT)(-0.39*N3);  // choose mode to check
+  INT nt1 = (INT)(0.37*N1), nt2 = (INT)(0.26*N2), nt3 = (INT)(-0.39*N3);  // choose mode to check
   dcomplex Ft = {0,0}, J = ima*(double)isign;
-  for (BIGINT j=0; j<M; ++j)
+  for (INT j=0; j<M; ++j)
     Ft += c[j] * exp(J*(nt1*x[j]+nt2*y[j]+nt3*z[j]));   // crude direct
   Ft /= M;
   // index in complex F as 1d array...
-  BIGINT it = N1/2+nt1 + N1*(N2/2+nt2) + N1*N2*(N3/2+nt3);
-  printf("one mode: rel err in F[%ld,%ld,%ld] is %.3g\n",nt1,nt2,nt3,
+  INT it = N1/2+nt1 + N1*(N2/2+nt2) + N1*N2*(N3/2+nt3);
+  printf("one mode: rel err in F[%ld,%ld,%ld] is %.3g\n",(INT64)nt1,(INT64)nt2,(INT64)nt3,
 	 abs(Ft-F[it])/infnorm(N,F));
   if (M*N<=BIGPROB) {                   // also check vs full direct eval
     dcomplex* Ft = (dcomplex*)malloc(sizeof(dcomplex)*N);
@@ -88,7 +88,7 @@ int main(int argc, char* argv[])
   }
 
   printf("test 3d type-2:\n"); // -------------- type 2
-  for (BIGINT m=0; m<N; ++m) F[m] = crandm11();
+  for (INT m=0; m<N; ++m) F[m] = crandm11();
   timer.restart();
   ier = finufft3d2(M,x,y,z,c,isign,tol,N1,N2,N3,F,opts);
   ti=timer.elapsedsec();
@@ -97,16 +97,16 @@ int main(int argc, char* argv[])
     exit(ier);
   } else
     printf("     (%ld,%ld,%ld) modes to %ld NU pts in %.3g s \t%.3g NU pts/s\n",
-	   N1,N2,N3,M,ti,M/ti);
+	   (INT64)N1,(INT64)N2,(INT64)N3,(INT64)M,ti,M/ti);
 
-  BIGINT jt = M/2;          // check arbitrary choice of one targ pt
+  INT jt = M/2;          // check arbitrary choice of one targ pt
   dcomplex ct = {0,0};
-  BIGINT m=0;
-  for (BIGINT m3=-(N3/2); m3<=(N3-1)/2; ++m3)   // loop in F order
-    for (BIGINT m2=-(N2/2); m2<=(N2-1)/2; ++m2)
-      for (BIGINT m1=-(N1/2); m1<=(N1-1)/2; ++m1)
+  INT m=0;
+  for (INT m3=-(N3/2); m3<=(N3-1)/2; ++m3)   // loop in F order
+    for (INT m2=-(N2/2); m2<=(N2-1)/2; ++m2)
+      for (INT m1=-(N1/2); m1<=(N1-1)/2; ++m1)
 	ct += F[m++] * exp(J*(m1*x[jt] + m2*y[jt] + m3*z[jt]));   // direct
-  printf("one targ: rel err in c[%ld] is %.3g\n",jt,abs(ct-c[jt])/infnorm(M,c));
+  printf("one targ: rel err in c[%ld] is %.3g\n",(INT64)jt,abs(ct-c[jt])/infnorm(M,c));
   if (M*N<=BIGPROB) {                  // also full direct eval
     dcomplex* ct = (dcomplex*)malloc(sizeof(dcomplex)*M);
     dirft3d2(M,x,y,z,ct,isign,N1,N2,N3,F);
@@ -116,7 +116,7 @@ int main(int argc, char* argv[])
 
   printf("test 3d type-3:\n"); // -------------- type 3
   // reuse the strengths c, interpret N as number of targs:
-  for (BIGINT j=0; j<M; ++j) {
+  for (INT j=0; j<M; ++j) {
     x[j] = 2.0 + M_PI*randm11();      // new x_j srcs, offset from origin
     y[j] = -3.0 + M_PI*randm11();     // " y_j
     z[j] = 1.0 + M_PI*randm11();     // " z_j
@@ -127,7 +127,7 @@ int main(int argc, char* argv[])
   double S1 = (double)N1/2;                   // choose freq range sim to type 1
   double S2 = (double)N2/2;
   double S3 = (double)N3/2;
-  for (BIGINT k=0; k<N; ++k) {
+  for (INT k=0; k<N; ++k) {
     s[k] = S1*(1.7 + randm11());    //S*(1.7 + k/(double)N); // offset the freqs
     t[k] = S2*(-0.5 + randm11());
     u[k] = S3*(0.9 + randm11());
@@ -139,13 +139,13 @@ int main(int argc, char* argv[])
     printf("error (ier=%d)!\n",ier);
     exit(ier);
   } else
-    printf("\t%ld NU to %ld NU in %.3g s   %.3g srcs/s, %.3g targs/s\n",M,N,ti,M/ti,N/ti);
+    printf("\t%ld NU to %ld NU in %.3g s   %.3g srcs/s, %.3g targs/s\n",(INT64)M,(INT64)N,ti,M/ti,N/ti);
 
-  BIGINT kt = N/2;          // check arbitrary choice of one targ pt
+  INT kt = N/2;          // check arbitrary choice of one targ pt
   Ft = {0,0};
-  for (BIGINT j=0;j<M;++j)
+  for (INT j=0;j<M;++j)
     Ft += c[j] * exp(ima*(double)isign*(s[kt]*x[j] + t[kt]*y[j] + u[kt]*z[j]));
-  printf("one targ: rel err in F[%ld] is %.3g\n",kt,abs(Ft-F[kt])/infnorm(N,F));
+  printf("one targ: rel err in F[%ld] is %.3g\n",(INT64)kt,abs(Ft-F[kt])/infnorm(N,F));
   if (M*N<=BIGPROB) {                  // also full direct eval
     dcomplex* Ft = (dcomplex*)malloc(sizeof(dcomplex)*N);
     dirft3d3(M,x,y,z,c,isign,N,s,t,u,Ft);       // writes to F
